requirements.txt
certifi==2020.12.5
chardet==4.0.0
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.2.2
future==0.18.2
gunicorn==20.0.4
idna==2.10
line-bot-sdk==1.18.0
psycopg2==2.8.6
pytz==2020.5
requests==2.25.1
sqlparse==0.4.1
static3==0.7.0
urllib3==1.26.2
beautifulsoup4 
pylint 
selenium==3.8.0
gspread
oauth2client
cdifflib
Flask==1.1.1
APScheduler==3.6.1
appnope==0.1.0
decorator==4.0.4
google-api-python-client
DateTime
PyYAML
pytz==2020.5
liffpy